requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.5
filelock==3.19.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
hf-xet==1.4.2
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.18.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pandas==2.3.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
protobuf==5.29.6
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pytest==8.3.5
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.3
rich==13.9.4
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.37.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
